Swift 4 + performance improvements (#8)
We ran into some serious performance improvements while using DoctorPretty for snapshot testing, so I wanted to see if there was anything I could do to fix it. I optimized a few small things (like using reduce(into:) with mutation instead of the pure version), but the real bottleneck is with the best and fits functions, which are recursive. I was able to rewrite fits as a plain loop, and that improved performance quite a bit, and we'd probably get huge improvements if we could rewrite best.